code updates with load balancer

+++ attendance-event-store/src/main/java/com/attendance/event/store/system/model/EmployeeAttedance.java
@@ -7,7 +7,7 @@ import java.sql.Date;
 @Data
 public class EmployeeAttedance {
 
-    private long employeeId;
+    private int employeeId;
     private String attendanceStatus;
     private Date date;
     private String totalTime;
